Fix Image show error

+++ app/src/main/java/me/ykrank/s1next/widget/glide/transformations/FitOutWidthBitmapTransformation.java
@@ -24,7 +24,7 @@ public final class FitOutWidthBitmapTransformation extends BitmapTransformation 
         float maxTextureSize = GlMaxTextureCalculator.getInstance().getGlMaxTextureSize();
         float textureSizeMultiplier = Math.min(maxTextureSize / resWidth, maxTextureSize / resHeight);
 
-        float fitOutSizeMultiplier = outWidth / resWidth;
+        float fitOutSizeMultiplier = (float) outWidth / resWidth;
         float sizeMultiplier = Math.min(textureSizeMultiplier, fitOutSizeMultiplier);
 
         if (sizeMultiplier < 1) {
